RAYSA-66 separate DB and models from each other
