+++ Flask Web App/app/models/db.py
@@ -1,0 +1,29 @@
+from os import name
+from flask_sqlalchemy import SQLAlchemy
+import datetime
+
+from flask_user import UserMixin
+from flask_user.forms import RegisterForm
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, validators
+db = SQLAlchemy()
+
+class BaseModel(db.Model):
+    """Base data model for all objects"""
+    __abstract__ = True
+    # def __init__(self, *args):
+    #     super().__init__(*args)
+    def __repr__(self):
+        """Define a base way to print models"""
+        return '%s(%s)' % (self.__class__.__name__, {
+            column: value
+            for column, value in self._to_dict().items()
+        })
+    def json(self):
+        """
+        Define a base way to jsonify models, dealing with datetime objects
+        """
+        return {
+            column: value if not isinstance(value, datetime.date) else value.strftime('%Y-%m-%d')
+            for column, value in self._to_dict().items()
+        }
